2010-04-27  Ozkan Sezer  <sezeroz@gmail.com>
	* dirent.h (DIR): Change dd_handle from long to intptr_t type for win64
	compatibility.
	(_WDIR): Likewise.



git-svn-id: svn+ssh://svn.code.sf.net/p/mingw-w64/code/trunk@2265 4407c894-4637-0410-b4f5-ada5f102cad1

+++ mingw-w64-headers/crt/dirent.h
@@ -44,7 +44,7 @@ extern "C" {
     struct dirent		dd_dir;
 
     /* _findnext handle */
-    long			dd_handle;
+    intptr_t			dd_handle;
 
     /*
     * Status of search:
@@ -91,7 +91,7 @@ extern "C" {
     struct _wdirent		dd_dir;
 
     /* _findnext handle */
-    long			dd_handle;
+    intptr_t			dd_handle;
 
     /*
     * Status of search:
